Make it possible for a client to specify a name for display.
This allows users to more easily know what they're logging in to.
There are two cases where users will hit this page:
- They're logging in for the first time
- They've just logged out of a client (which then refreshed, and
    sent them here as a result of not having their own logout page)
In both of these cases (though the second particularly), it's very
useful for the user to know where they came from.

TODO: propogate this value throughout so we can be nicer to the user
once they've logged in, too.

Change-Id: I5b821aaf781578eac3cdd27863486e625dc9a3e8